+++ src/EarlyStopping/__init__.py
@@ -1,3 +1,4 @@
 from .example import add_one   # '.' needed since python3
 from .L2Boost import L2Boost
-from .FilterEstimator import FilterEstimator, EstimationMethod
+from .Landweber import Landweber
+
